Added support for prefix '_' and '$' in AccessorsInfo.
The '@Accessors' documentation says: 'For characters which are letters,
the character following the prefix must not be a lowercase letter ...'
(http://projectlombok.org/features/experimental/Accessors.html).

Both characters '_' and '$' are valid prefixes of Java identifiers. The
Plugin should not require the first letter after these prefixes to be
uppercase.