Removed the dependency on lombok version 1.16.20, which was used for the experimental versions of the @Builder and @Value annotations.
@Builder was promoted in lombok from experimental to the main package in v1.16.0 on January 26th, 2015, nearly 4 years ago.
@Value was promoted in lombok from experimental to the main package in v0.12.0 on July 16th, 2013, over 5 years ago.

The experimental versions of @Builder and @Value have been replaced with the main package versions, or otherwise removed where appropriate.

The chain and fluent options to @Builder are now the default, so the BuilderChainAndFluent test was removed.
The ValueExperimental test no longer seems relevant so it was also removed.
Other tests were updated as necessary.